requirements.txt
allure_python_commons==2.9.45
google_crc32c==1.3.0
ocpp==0.11.0
pytest==6.2.5
python_dateutil==2.8.2
PyYAML==6.0
websockets==10.1
